refactor(domain): extract auth/ssh/project-factory logic from facade
The facade was historically the single stable import boundary for terok,
but lib.api was introduced (PR #904) to replace it for the TUI and the
CLI was the last consumer keeping it alive.  Its internal logic had no
real home — carving it out is the first step toward retiring the module:

- lib/domain/ssh.py (new): SSH provisioning workflow — provision_ssh_key,
  register_ssh_key, summarize_ssh_init, project_needs_key_registration,
  maybe_pause_for_ssh_key_registration.
- lib/domain/auth.py (new): host-wide and project-scoped auth flow —
  authenticate plus _resolve_host_auth_image's three-way image-build
  prompt.
- lib/domain/project.py: project factory functions get_project,
  list_projects, derive_project, _share_ssh_key_assignments, and
  project_image_exists land next to the Project aggregate they construct.

facade.py is now a thin re-export shim so existing CLI imports keep
working until the next commit migrates them to terok.lib.api.

Tests follow the code: test_facade.py patches retarget each function's
defining module (mock.patch resolves names where they are looked up at
call time, not where they were imported from), and test_ssh.py,
test_projects.py, test_cli_workflows.py patches updated to match.

tach.toml and .importlinter declare the new modules and the (still
transitional) allowed-importer lists for the sibling-wheel boundaries.
