+++ tests/unit/lib/domain/test_facade.py
@@ -1,13 +1,24 @@
 # SPDX-FileCopyrightText: 2026 Jiri Vyskocil
 # SPDX-License-Identifier: Apache-2.0
 
-"""Tests for the domain.facade thin-wrapper factories."""
+"""Tests for the domain.facade re-export shim's underlying logic.
+
+The functions live in ``terok.lib.domain.{project,ssh,auth}``; facade
+re-exports them for the transitional period.  These tests patch each
+function's *defining* module, since that is where ``mock.patch`` resolves
+the names actually referenced at call time.
+"""
 
 from __future__ import annotations
 
+from contextlib import contextmanager
 from unittest.mock import MagicMock, patch
 
 import pytest
+
+# ---------------------------------------------------------------------------
+# Project factory functions (defined in terok.lib.domain.project)
+# ---------------------------------------------------------------------------
 
 
 class TestGetProject:
@@ -19,7 +30,7 @@ class TestGetProject:
 
         fake_cfg = MagicMock()
         fake_cfg.id = "myproj"
-        with patch("terok.lib.domain.facade.load_project", return_value=fake_cfg) as loader:
+        with patch("terok.lib.domain.project.load_project", return_value=fake_cfg) as loader:
             result = facade.get_project("myproj")
         loader.assert_called_once_with("myproj")
         assert isinstance(result, Project)
@@ -33,7 +44,7 @@ class TestListProjects:
         from terok.lib.domain.project import Project
 
         a, b = MagicMock(id="a"), MagicMock(id="b")
-        with patch("terok.lib.core.projects.list_projects", return_value=[a, b]) as lister:
+        with patch("terok.lib.domain.project._list_projects", return_value=[a, b]) as lister:
             result = facade.list_projects()
         lister.assert_called_once()
         assert len(result) == 2
@@ -42,7 +53,7 @@ class TestListProjects:
     def test_empty_list_returns_empty(self) -> None:
         from terok.lib.domain import facade
 
-        with patch("terok.lib.core.projects.list_projects", return_value=[]):
+        with patch("terok.lib.domain.project._list_projects", return_value=[]):
             assert facade.list_projects() == []
 
 
@@ -55,9 +66,9 @@ class TestDeriveProject:
 
         derived_cfg = MagicMock(id="derived")
         with (
-            patch("terok.lib.domain.facade._derive_project") as derive,
-            patch("terok.lib.domain.facade._share_ssh_key_assignments") as share,
-            patch("terok.lib.domain.facade.load_project", return_value=derived_cfg) as loader,
+            patch("terok.lib.domain.project._derive_project") as derive,
+            patch("terok.lib.domain.project._share_ssh_key_assignments") as share,
+            patch("terok.lib.domain.project.load_project", return_value=derived_cfg) as loader,
         ):
             result = facade.derive_project("source", "derived")
         derive.assert_called_once_with("source", "derived")
@@ -66,29 +77,38 @@ class TestDeriveProject:
         assert isinstance(result, Project)
 
 
-def _patch_vault_db(db):
-    """Patch ``facade.vault_db`` to yield *db* — returns the ``patch`` context."""
-    from contextlib import contextmanager
+# ---------------------------------------------------------------------------
+# Vault-DB patch helpers, one per consumer module
+# ---------------------------------------------------------------------------
+
+
+def _patch_vault_db(db, *, module: str):
+    """Patch ``vault_db`` in the given consumer module to yield *db*."""
 
     @contextmanager
     def _cm():
         yield db
 
-    return patch("terok.lib.domain.facade.vault_db", _cm)
+    return patch(f"terok.lib.domain.{module}.vault_db", _cm)
+
+
+# ---------------------------------------------------------------------------
+# SSH workflow (defined in terok.lib.domain.ssh)
+# ---------------------------------------------------------------------------
 
 
 class TestShareSshKeyAssignments:
     """Copy every SSH key assignment from the source scope to the new scope."""
 
     def test_delegates_to_db_assign_for_each_row(self) -> None:
-        from terok.lib.domain import facade
+        from terok.lib.domain import project
 
         row_a = MagicMock(id=1)
         row_b = MagicMock(id=2)
         db = MagicMock()
         db.list_ssh_keys_for_scope.return_value = [row_a, row_b]
-        with _patch_vault_db(db):
-            facade._share_ssh_key_assignments("src", "new")
+        with _patch_vault_db(db, module="project"):
+            project._share_ssh_key_assignments("src", "new")
         db.list_ssh_keys_for_scope.assert_called_once_with("src")
         assert db.assign_ssh_key.call_args_list == [
             (("new", 1),),
@@ -96,12 +116,12 @@ class TestShareSshKeyAssignments:
         ]
 
     def test_silent_noop_when_source_has_no_keys(self) -> None:
-        from terok.lib.domain import facade
+        from terok.lib.domain import project
 
         db = MagicMock()
         db.list_ssh_keys_for_scope.return_value = []
-        with _patch_vault_db(db):
-            facade._share_ssh_key_assignments("src", "new")
+        with _patch_vault_db(db, module="project"):
+            project._share_ssh_key_assignments("src", "new")
         db.assign_ssh_key.assert_not_called()
 
 
@@ -112,7 +132,7 @@ class TestRegisterSshKey:
         from terok.lib.domain import facade
 
         db = MagicMock()
-        with _patch_vault_db(db):
+        with _patch_vault_db(db, module="ssh"):
             facade.register_ssh_key("myproj", 42)
         db.assign_ssh_key.assert_called_once_with("myproj", 42)
 
@@ -137,9 +157,9 @@ class TestProvisionSshKey:
 
         db = MagicMock()
         with (
-            patch("terok.lib.domain.facade.load_project", return_value=MagicMock(id="myproj")),
+            patch("terok.lib.domain.ssh.load_project", return_value=MagicMock(id="myproj")),
             patch("terok.lib.domain.project.make_ssh_manager", return_value=ssh_manager),
-            _patch_vault_db(db),
+            _patch_vault_db(db, module="ssh"),
         ):
             result = facade.provision_ssh_key("myproj", key_type="ed25519", force=True)
 
@@ -179,7 +199,7 @@ class TestMaybePauseForSshKeyRegistration:
 
         project = MagicMock(upstream_url="git@example.com:org/repo.git")
         with (
-            patch("terok.lib.domain.facade.load_project", return_value=project),
+            patch("terok.lib.domain.ssh.load_project", return_value=project),
             patch("builtins.input", return_value=""),
         ):
             facade.maybe_pause_for_ssh_key_registration("myproj")
@@ -190,7 +210,7 @@ class TestMaybePauseForSshKeyRegistration:
 
         project = MagicMock(upstream_url="ssh://git@example.com/org/repo.git")
         with (
-            patch("terok.lib.domain.facade.load_project", return_value=project),
+            patch("terok.lib.domain.ssh.load_project", return_value=project),
             patch("builtins.input", return_value=""),
         ):
             facade.maybe_pause_for_ssh_key_registration("myproj")
@@ -200,7 +220,7 @@ class TestMaybePauseForSshKeyRegistration:
         from terok.lib.domain import facade
 
         project = MagicMock(upstream_url="https://github.com/org/repo.git")
-        with patch("terok.lib.domain.facade.load_project", return_value=project):
+        with patch("terok.lib.domain.ssh.load_project", return_value=project):
             facade.maybe_pause_for_ssh_key_registration("myproj")
         assert "ACTION REQUIRED" not in capsys.readouterr().out
 
@@ -208,9 +228,14 @@ class TestMaybePauseForSshKeyRegistration:
         from terok.lib.domain import facade
 
         project = MagicMock(upstream_url=None)
-        with patch("terok.lib.domain.facade.load_project", return_value=project):
+        with patch("terok.lib.domain.ssh.load_project", return_value=project):
             facade.maybe_pause_for_ssh_key_registration("myproj")
         assert "ACTION REQUIRED" not in capsys.readouterr().out
+
+
+# ---------------------------------------------------------------------------
+# Auth flow (defined in terok.lib.domain.auth)
+# ---------------------------------------------------------------------------
 
 
 class TestAuthenticate:
@@ -222,15 +247,15 @@ class TestAuthenticate:
 
         # sandbox_live_mounts_dir and the expose-token predicates are
         # lazy-imported inside the function body, so patching happens
-        # at their definition modules rather than on the facade.
+        # at their definition modules rather than on the auth module.
         with (
             patch(
-                "terok.lib.domain.facade.project_cli_image", return_value="terok-p1:latest"
+                "terok.lib.domain.auth.project_cli_image", return_value="terok-p1:latest"
             ) as mock_l2,
             patch("terok.lib.core.config.sandbox_live_mounts_dir", return_value="/mnt"),
             patch("terok.lib.core.config.is_claude_oauth_exposed", return_value=False),
             patch("terok.lib.core.config.is_codex_oauth_exposed", return_value=False),
-            patch("terok.lib.domain.facade._authenticate_raw") as mock_auth,
+            patch("terok.lib.domain.auth._authenticate_raw") as mock_auth,
         ):
             facade.authenticate("claude", project_id="p1")
 
@@ -248,13 +273,13 @@ class TestAuthenticate:
 
         with (
             patch(
-                "terok.lib.domain.facade._resolve_host_auth_image",
+                "terok.lib.domain.auth._resolve_host_auth_image",
                 return_value="terok-l1-cli:ubuntu-24.04",
             ) as mock_resolve,
             patch("terok.lib.core.config.sandbox_live_mounts_dir", return_value="/mnt"),
             patch("terok.lib.core.config.is_claude_oauth_exposed", return_value=False),
             patch("terok.lib.core.config.is_codex_oauth_exposed", return_value=False),
-            patch("terok.lib.domain.facade._authenticate_raw") as mock_auth,
+            patch("terok.lib.domain.auth._authenticate_raw") as mock_auth,
         ):
             facade.authenticate("claude")
 
@@ -276,13 +301,13 @@ class TestAuthenticate:
 
         with (
             patch(
-                "terok.lib.domain.facade._resolve_host_auth_image",
+                "terok.lib.domain.auth._resolve_host_auth_image",
                 return_value="terok-l1-cli:ubuntu-24.04",
             ),
             patch("terok.lib.core.config.sandbox_live_mounts_dir", return_value="/mnt"),
             patch("terok.lib.core.config.is_claude_oauth_exposed", return_value=False),
             patch("terok.lib.core.config.is_codex_oauth_exposed", return_value=True),
-            patch("terok.lib.domain.facade._authenticate_raw") as mock_auth,
+            patch("terok.lib.domain.auth._authenticate_raw") as mock_auth,
         ):
             facade.authenticate("codex")
 
@@ -294,13 +319,13 @@ class TestAuthenticate:
 
         with (
             patch(
-                "terok.lib.domain.facade._resolve_host_auth_image",
+                "terok.lib.domain.auth._resolve_host_auth_image",
                 return_value="terok-l1-cli:ubuntu-24.04",
             ),
             patch("terok.lib.core.config.sandbox_live_mounts_dir", return_value="/mnt"),
             patch("terok.lib.core.config.is_claude_oauth_exposed", return_value=False),
             patch("terok.lib.core.config.is_codex_oauth_exposed", return_value=True),
-            patch("terok.lib.domain.facade._authenticate_raw") as mock_auth,
+            patch("terok.lib.domain.auth._authenticate_raw") as mock_auth,
         ):
             facade.authenticate("claude")
 
@@ -313,14 +338,14 @@ class TestAuthenticate:
 
         with (
             patch(
-                "terok.lib.domain.facade._resolve_host_auth_image",
+                "terok.lib.domain.auth._resolve_host_auth_image",
                 return_value="terok-l1-cli:ubuntu-24.04",
             ),
             patch("terok.lib.core.config.sandbox_live_mounts_dir", return_value="/mnt"),
             patch("terok.lib.core.config.is_claude_oauth_exposed", return_value=False),
             patch("terok.lib.core.config.is_codex_oauth_exposed", return_value=False),
             patch("terok.lib.core.config.is_oauth_enabled_for", return_value=False) as mock_gate,
-            patch("terok.lib.domain.facade._authenticate_raw") as mock_auth,
+            patch("terok.lib.domain.auth._authenticate_raw") as mock_auth,
         ):
             facade.authenticate("codex")
 
@@ -333,14 +358,14 @@ class TestAuthenticate:
 
         with (
             patch(
-                "terok.lib.domain.facade._resolve_host_auth_image",
+                "terok.lib.domain.auth._resolve_host_auth_image",
                 return_value="terok-l1-cli:ubuntu-24.04",
             ),
             patch("terok.lib.core.config.sandbox_live_mounts_dir", return_value="/mnt"),
             patch("terok.lib.core.config.is_claude_oauth_exposed", return_value=False),
             patch("terok.lib.core.config.is_codex_oauth_exposed", return_value=False),
             patch("terok.lib.core.config.is_oauth_enabled_for", return_value=True),
-            patch("terok.lib.domain.facade._authenticate_raw") as mock_auth,
+            patch("terok.lib.domain.auth._authenticate_raw") as mock_auth,
         ):
             facade.authenticate("claude")
 
@@ -352,21 +377,21 @@ class TestResolveHostAuthImage:
 
     def test_prefers_default_alias_when_provider_label_present(self) -> None:
         """Default alias wins iff the OCI label confirms it has the provider."""
-        from terok.lib.domain import facade
+        from terok.lib.domain import auth
 
         def exists(tag: str) -> bool:
             return tag == "terok-l1-cli:ubuntu-24.04"
 
         with (
-            patch("terok.lib.domain.facade.image_exists", side_effect=exists),
+            patch("terok.lib.domain.auth.image_exists", side_effect=exists),
             patch("terok_executor.image_agents", return_value={"claude", "codex", "gh"}),
         ):
-            image = facade._resolve_host_auth_image("claude")
+            image = auth._resolve_host_auth_image("claude")
         assert image == "terok-l1-cli:ubuntu-24.04"
 
     def test_alias_present_but_label_missing_provider_falls_through(self) -> None:
         """Stale alias from a partial build must not silently route auth at it."""
-        from terok.lib.domain import facade
+        from terok.lib.domain import auth
 
         # Stale alias points at a claude-only image; per-agent for codex
         # was previously built too.  Asking for codex must skip the alias
@@ -378,57 +403,57 @@ class TestResolveHostAuthImage:
             }
 
         with (
-            patch("terok.lib.domain.facade.image_exists", side_effect=exists),
+            patch("terok.lib.domain.auth.image_exists", side_effect=exists),
             patch("terok_executor.image_agents", return_value={"claude"}),
         ):
-            image = facade._resolve_host_auth_image("codex")
+            image = auth._resolve_host_auth_image("codex")
         assert image == "terok-l1-cli:ubuntu-24.04-codex"
 
     def test_falls_back_to_per_agent_l1(self) -> None:
-        from terok.lib.domain import facade
+        from terok.lib.domain import auth
 
         def exists(tag: str) -> bool:
             return tag == "terok-l1-cli:ubuntu-24.04-claude"
 
         with (
-            patch("terok.lib.domain.facade.image_exists", side_effect=exists),
+            patch("terok.lib.domain.auth.image_exists", side_effect=exists),
             patch("terok_executor.image_agents", return_value=set()),
         ):
-            image = facade._resolve_host_auth_image("claude")
+            image = auth._resolve_host_auth_image("claude")
         assert image == "terok-l1-cli:ubuntu-24.04-claude"
 
     def test_api_key_only_provider_skips_build_when_missing(self) -> None:
         """API-key-only providers never launch a container; any tag is fine."""
-        from terok.lib.domain import facade
+        from terok.lib.domain import auth
 
         with (
-            patch("terok.lib.domain.facade.image_exists", return_value=False),
+            patch("terok.lib.domain.auth.image_exists", return_value=False),
             patch("terok_executor.image_agents", return_value=set()),
         ):
             # sonar is api-key-only — no prompt, no build, just a tag.
-            image = facade._resolve_host_auth_image("sonar")
+            image = auth._resolve_host_auth_image("sonar")
         assert image.startswith("terok-l1-cli:")
 
     def test_oauth_provider_exits_on_non_tty_without_image(self) -> None:
         """Non-TTY OAuth run with no image exits with a build hint, no prompt."""
-        from terok.lib.domain import facade
+        from terok.lib.domain import auth
 
         with (
-            patch("terok.lib.domain.facade.image_exists", return_value=False),
+            patch("terok.lib.domain.auth.image_exists", return_value=False),
             patch("terok_executor.image_agents", return_value=set()),
             patch("sys.stdin.isatty", return_value=False),
             patch("sys.stdout.isatty", return_value=False),
             pytest.raises(SystemExit) as exc,
         ):
-            facade._resolve_host_auth_image("claude")
+            auth._resolve_host_auth_image("claude")
         assert "terok project build" in str(exc.value) or "terok executor" in str(exc.value)
 
     def test_y_choice_builds_full_default(self) -> None:
         """[Y] (default) routes to ensure_default_l1 with the user's agent set."""
-        from terok.lib.domain import facade
+        from terok.lib.domain import auth
 
         with (
-            patch("terok.lib.domain.facade.image_exists", return_value=False),
+            patch("terok.lib.domain.auth.image_exists", return_value=False),
             patch("terok_executor.image_agents", return_value=set()),
             patch("sys.stdin.isatty", return_value=True),
             patch("sys.stdout.isatty", return_value=True),
@@ -444,17 +469,17 @@ class TestResolveHostAuthImage:
             ) as mock_ensure,
             patch("terok_executor.build_base_images") as mock_per_agent,
         ):
-            tag = facade._resolve_host_auth_image("claude")
+            tag = auth._resolve_host_auth_image("claude")
         assert tag == "terok-l1-cli:ubuntu-24.04"
         mock_ensure.assert_called_once()
         mock_per_agent.assert_not_called()
 
     def test_one_choice_builds_minimal_per_agent(self) -> None:
         """[1] builds only the per-agent L1, not the full default."""
-        from terok.lib.domain import facade
+        from terok.lib.domain import auth
 
         with (
-            patch("terok.lib.domain.facade.image_exists", return_value=False),
+            patch("terok.lib.domain.auth.image_exists", return_value=False),
             patch("terok_executor.image_agents", return_value=set()),
             patch("sys.stdin.isatty", return_value=True),
             patch("sys.stdout.isatty", return_value=True),
@@ -467,7 +492,7 @@ class TestResolveHostAuthImage:
             patch("terok_executor.ensure_default_l1") as mock_ensure,
             patch("terok_executor.build_base_images") as mock_per_agent,
         ):
-            tag = facade._resolve_host_auth_image("codex")
+            tag = auth._resolve_host_auth_image("codex")
         assert tag == "terok-l1-cli:ubuntu-24.04-codex"
         mock_per_agent.assert_called_once()
         # Per-agent build with just the requested provider, no alias retag
@@ -479,10 +504,10 @@ class TestResolveHostAuthImage:
 
     def test_n_choice_aborts(self) -> None:
         """[n] exits with the build-hint message."""
-        from terok.lib.domain import facade
+        from terok.lib.domain import auth
 
         with (
-            patch("terok.lib.domain.facade.image_exists", return_value=False),
+            patch("terok.lib.domain.auth.image_exists", return_value=False),
             patch("terok_executor.image_agents", return_value=set()),
             patch("sys.stdin.isatty", return_value=True),
             patch("sys.stdout.isatty", return_value=True),
@@ -494,18 +519,18 @@ class TestResolveHostAuthImage:
             patch("builtins.input", return_value="n"),
             pytest.raises(SystemExit) as exc,
         ):
-            facade._resolve_host_auth_image("codex")
+            auth._resolve_host_auth_image("codex")
         assert "terok project build" in str(exc.value) or "terok executor" in str(exc.value)
 
     def test_uses_configured_base_image(self) -> None:
         """Honors ``image.base_image`` from config — not hardcoded ubuntu."""
-        from terok.lib.domain import facade
+        from terok.lib.domain import auth
 
         def exists(tag: str) -> bool:
             return tag == "terok-l1-cli:fedora-43"
 
         with (
-            patch("terok.lib.domain.facade.image_exists", side_effect=exists),
+            patch("terok.lib.domain.auth.image_exists", side_effect=exists),
             patch("terok_executor.image_agents", return_value={"claude", "codex", "gh"}),
             patch(
                 "terok.lib.core.config.get_global_image_base_image",
@@ -513,5 +538,5 @@ class TestResolveHostAuthImage:
             ),
             patch("terok.lib.core.config.get_global_image_agents", return_value="all"),
         ):
-            image = facade._resolve_host_auth_image("claude")
+            image = auth._resolve_host_auth_image("claude")
         assert image == "terok-l1-cli:fedora-43"
