+++ tests/unit/lib/test_ssh.py
@@ -10,21 +10,21 @@ from unittest.mock import MagicMock, patch
 
 
 def _patch_vault_db(db):
-    """Patch ``facade.vault_db`` to yield *db*."""
+    """Patch the SSH module's ``vault_db`` alias to yield *db*."""
 
     @contextmanager
     def _cm():
         yield db
 
-    return patch("terok.lib.domain.facade.vault_db", _cm)
+    return patch("terok.lib.domain.ssh.vault_db", _cm)
 
 
 class TestRegisterSshKey:
-    """Tests for ``register_ssh_key`` in the domain facade."""
+    """Tests for ``register_ssh_key`` in the SSH workflow module."""
 
     def test_assigns_key_to_scope(self) -> None:
         """register_ssh_key delegates to ``CredentialDB.assign_ssh_key``."""
-        from terok.lib.domain.facade import register_ssh_key
+        from terok.lib.domain.ssh import register_ssh_key
 
         db = MagicMock()
         with _patch_vault_db(db):
@@ -35,7 +35,7 @@ class TestRegisterSshKey:
         """Errors from the DB layer propagate (no silent swallowing)."""
         import pytest
 
-        from terok.lib.domain.facade import register_ssh_key
+        from terok.lib.domain.ssh import register_ssh_key
 
         db = MagicMock()
         db.assign_ssh_key.side_effect = RuntimeError("disk full")
